+++ hugolib/site.go
@@ -500,6 +500,15 @@ func (s *Site) assembleSections() {
 
 	for k := range s.Sections {
 		s.Sections[k].Sort()
+
+		for i, wp := range s.Sections[k] {
+			if i > 0 {
+				wp.Page.PrevInSection = s.Sections[k][i - 1].Page;
+			}
+			if i < len(s.Sections[k]) - 1 {
+				wp.Page.NextInSection = s.Sections[k][i + 1].Page;
+			}
+		}
 	}
 }